requirements.txt
Flask==1.1.2
pymongo==3.8.0
boto3===1.13.19
gevent
requests
pprint
pillow
reportlab